requirements.txt
Flask
Flask-SQLAlchemy
requests
flask-wtf
Flask-APScheduler
